[als] make ALS return more sane values. Contributes: JB#31875
Try to make the ALS output some more sane values. The ALS was too
quickly falling back to 0, this seems to have to do with the division
by a huge number followed by a multiplication. As a small value will
soon be reduced to 0 that way. By applying values used by the driver
by default this behaviour seems to be improved and we do not get zero
values from the ALS so quickly anymore.

Signed-off-by: Philippe De Swert <philippe.deswert@jollamobile.com>