input: misc: qpnp-vibra: Block suspend in prepare if vibra is active.
The vibra may be left in a running state during suspend transition
if there is a ramp phase going on in ff-memless side as the
ff-memless core does not handle suspend in any way and may set
speed of the driver after the driver has suspended. To fix this,
disallow suspend in the driver side when vibra is on. This is done in
earliest possible phase (prepare) to reduce transition latency.

[vibra] Block suspend in prepare if vibra is active. Fixes JB#33643

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jolla.com>