input: misc: qpnp-vibra: Do not lock mutex in timer function
Timer functions must not block. The qpnp_vib_play() is called from
a timer function and locks a mutex. The mutex is not used anywhere
else. Calls to qpnp_vib_play() from ff_memless side are made while
holding a spinlock.

Remove the unnecessary and potentially dangerous mutex.

Rearrange code in qpnp_vib_play() so that intermediate zero value
can't show up in vib->speed, and qpnp_vib_work() so that the value
read from vib->speed is used for the whole duration of the call.

[vibra] Do not lock mutex in timer function. Contributes to JB#33972

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>